# payment/payment/static/swagger.json
{"info": {"title": "StrongPay", "version": "1.0.0"}, "paths": {"/CreateCheckout": {"post": {"description": "Creates a checkout", "tags": ["Payment"], "requestBody": {"required": true, "content": {"application/x-www-form-urlencoded": {"schema": {"type": "object", "properties": {"AMOUNT": {"type": "double", "description": "Amount to be paid by the client"}, "MERCHANT": {"type": "string", "description": "Token that indentifies the merchant"}, "RETURN_URL": {"type": "string", "description": "URL to where the client in redirect if the payment is successful"}, "CANCEL_URL": {"type": "string", "description": "URL to where the client in redirect if the payment is cancelled"}}, "required": ["AMOUNT", "MERCHANT", "RETURN_URL", "CANCEL_URL"]}}}}, "responses": {"201": {"description": "A JSON containing a TOKEN that indentifies the Checkout", "content": {"application/json": {"schema": {"properties": {"TOKEN": {"type": "string"}}}}}}, "400": {"description": "A JSON containing a ERROR that indentifies the problem", "content": {"application/json": {"schema": {"properties": {"ERROR": {"type": "string"}}}}}}}}}, "/GetCheckoutDetails": {"get": {"description": "Gets checkout details such as who paid and billing address", "tags": ["Payment"], "parameters": [{"in": "path", "name": "checkout_token", "schema": {"type": "string"}, "required": true, "description": "Checkout token given by create_checkout"}], "responses": {"200": {"description": "A JSON containing buyer information and checkout status (not defined yet)", "content": {"application/json": {"schema": {"properties": {"NAME": {"type": "string"}, "BUYER_ID": {"type": "string"}}}}}}, "400": {"description": "A JSON containing a ERROR that indentifies the problem", "content": {"application/json": {"schema": {"properties": {"ERROR": {"type": "string"}}}}}}}}}, "/ExecuteCheckout": {"get": {"description": "Confirm and execute payment. Returns Success or Failure.", "tags": ["Payment"], "parameters": [{"in": "path", "name": "checkout_token", "schema": {"type": "string"}, "required": true, "description": "Checkout token given by create_checkout"}, {"in": "path", "name": "buyer_id", "schema": {"type": "string"}, "required": true, "description": "Buyer ID given by /GetCheckoutDetails"}], "responses": {"200": {"description": "A JSON containing result of payment", "content": {"application/json": {"schema": {"properties": {"SUCCESS": {"type": "boolean"}}}}}}, "400": {"description": "A JSON containing a ERROR that indentifies the problem", "content": {"application/json": {"schema": {"properties": {"ERROR": {"type": "string"}}}}}}}}}}, "tags": [], "openapi": "3.0.0", "components": {"schemas": {}, "parameters": {}}, "base_path": "http://127.0.0.1:5000/"}